Объединил тесты для удобства
