+++ js/data/answer.js
@@ -1,3 +1,5 @@
+import {initialState} from '../data';
+
 export const checkAnswer = (answers) => {
   const result = answers.every((answer) => {
     return answer.isRight;
@@ -6,4 +8,17 @@ export const checkAnswer = (answers) => {
     return true;
   }
   return false;
+};
+
+export const addPoint = (state) => {
+  const differenceTime = initialState.timer - state.timer;
+  const userStats = state.stats;
+  if (differenceTime < 10) {
+    userStats.push(`fast`);
+  }
+  if (differenceTime > 20) {
+    userStats.push(`slow`);
+  }
+  const newState = Object.assign({}, state, {stats: userStats});
+  return newState;
 };
